+++ backend/store/services/image-service/src/main/java/com/team/imageservice/service/api/ImageDeliveryService.java
@@ -1,4 +1,4 @@
-package com.team.imageservice.service;
+package com.team.imageservice.service.api;
 
 import com.team.imageservice.data.Image;
